requirements.txt
tensorflow
tensorflow-datasets
numpy
matplotlib
pillow
streamlit
